v2/analyzer/requirements.txt
openai>=1.0.0
assemblyai>=0.17.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
moviepy>=1.0.3
rapidfuzz>=3.0.0

